@@ -18,7 +18,7 @@
     "url": "https://mediasuite.co.nz"
   },
   "engines": {
-    "node": ">=4.2.0"
+    "node": ">=8.11.0"
   },
   "scripts": {
     "lint": "standard --verbose | snazzy",
@@ -46,19 +46,18 @@
     "download"
   ],
   "dependencies": {
-    "aws-sdk": "^2.2.28",
-    "pkgcloud": "^1.1.0",
-    "ramda": "^0.19.1",
-    "sharp": "^0.15.1"
+    "aws-sdk": "^2.450.0",
+    "pkgcloud": "^1.7.0",
+    "sharp": "^0.22.1"
   },
   "devDependencies": {
     "chai": "^3.4.1",
-    "mocha": "^3.0.2",
+    "mocha": "^6.1.4",
     "rewire": "^2.5.1",
-    "snazzy": "^2.0.1",
-    "standard": "^5.4.1",
-    "testem": "^1.10.3",
-    "xyz": "^0.5.0"
+    "snazzy": "^8.0.0",
+    "standard": "^12.0.1",
+    "testem": "^2.15.1",
+    "xyz": "^3.0.0"
   },
   "peerDependencies": {}
 }
