@@ -6,13 +6,13 @@ const expect = require('chai').expect
 const fs = require('fs')
 const cp = require('child_process')
 const rewire = require('rewire')
-let adapter = rewire('../adapters/amazon')
+const adapter = rewire('../adapters/amazon')
 
 const AWS = {
   S3: function () {
     this.upload = (params, cb) => {
       fs.writeFileSync(`${__dirname}/files/${params.Key}`, params.Body)
-      cb(null, {key: params.Key})
+      cb(null, { key: params.Key })
     }
     this.getObject = (params, cb) => {
       cb(null, fs.createReadStream(`${__dirname}/files/${params.Key}`))
@@ -35,23 +35,17 @@ describe('amazon adapter', () => {
     afterEach(() => {
       cp.execSync(`rm -rf ${__dirname}/files`)
     })
-    it('upload should successfully upload a file', (done) => {
+    it('upload should successfully upload a file', async () => {
       // given
-      let config = { provider: 'amazon', container: `${__dirname}/files` }
-      let client = adapter(config)
+      const config = { provider: 'amazon', container: `${__dirname}/files` }
+      const client = adapter(config)
 
       // when
-      let upload = client.upload('my-file.txt', new Buffer('asdasdasd'), {})
+      await client.upload('my-file.txt', Buffer.from('asdasdasd'), {})
 
       // then
-      upload.then(() => {
-        let file = fs.readFileSync(`${__dirname}/files/my-file.txt`, 'utf8')
-        expect(file).to.equal('asdasdasd')
-        done()
-      }).catch(() => {
-        expect(false).to.be.ok
-        done()
-      })
+      const file = fs.readFileSync(`${__dirname}/files/my-file.txt`, 'utf8')
+      expect(file).to.equal('asdasdasd')
     })
   })
   describe('downloading a file', () => {
@@ -67,44 +61,35 @@ describe('amazon adapter', () => {
 
     it('download should successfully download a file', (done) => {
       // given
-      let config = { provider: 'amazon', container: 'dummy' }
-      let client = adapter(config)
+      const config = { provider: 'amazon', container: 'dummy' }
+      const client = adapter(config)
 
       // when
-      let download = client.download('my-file.txt', {})
+      const download = client.download('my-file.txt', {})
 
       // then
       download.then(fileStream => {
         let data = ''
-        fileStream.on('data', chunk => data += chunk)
+        fileStream.on('data', chunk => { data += chunk })
         fileStream.on('end', () => {
           expect(data).to.equal('asdasdasd\n')
           done()
         })
-      }).catch(() => {
-        expect(false).to.be.ok
-        done()
       })
     })
   })
 
   describe('getting a signed url', () => {
-    it('should return a promise that resolves to a url', (done) => {
+    it('should return a promise that resolves to a url', async () => {
       // given
-      let config = { provider: 'amazon', container: 'dummy' }
-      let client = adapter(config)
+      const config = { provider: 'amazon', container: 'dummy' }
+      const client = adapter(config)
 
       // when
-      let url = client.getUrl('my-file.txt', {})
+      const url = await client.getUrl('my-file.txt', {})
 
       // then
-      url.then(url => {
-        expect(url).to.equal('http://sub.domain.tld/path/to/file')
-        done()
-      }).catch(() => {
-        expect(false).to.be.ok
-        done()
-      })
+      expect(url).to.equal('http://sub.domain.tld/path/to/file')
     })
   })
 })
